requirements.txt
httpx
fastapi
uvicorn[standard]
faiss-cpu
numpy
pillow
pytest
pytest-asyncio
